requirements.txt
fastapi
uvicorn
pymongo
scikit-learn 
openai 
python-dotenv
chromadb
sentence-transformers